tests: Avoid the chance of variable expansion to `/*`
It is dangerous to have in the code the possibility of an empty variable
to lead to the erasure of something like `/*`. Here the `:?` exits the
script if the parameter is null or unset

+++ tests/tools/run_test_concurrently.sh
@@ -71,7 +71,7 @@ make -C build/lizardfs -j$(nproc) install
 
 killall -9 lizardfs-tests || true
 mkdir -m 777 -p $TEST_OUTPUT_DIR
-rm -rf $TEST_OUTPUT_DIR/* || true
+rm -rf "${TEST_OUTPUT_DIR:?}"/* || true
 rm -rf /mnt/ramdisk/* || true
 
 FILTER=$(python3 $WORKSPACE/tests/tools/filter_tests.py --workspace $WORKSPACE --test_suite $TEST_SUITE \
